Compose BioCompute Object - Error Domain (v1.4.2)

The error domain can be used to determine what range of input returns outputs that are within the tolerance level defined in this subdomain and therefore can be used to optimize algorithm (domain definition).

compose_error_v1.4.2(empirical = NULL, algorithmic = NULL)

compose_error(empirical = NULL, algorithmic = NULL)

Arguments

empirical

Data frame. Variables include key and value. Each row is one item in the empirical error subdomain.

algorithmic

Data frame. Variables include key and value. Each row is one item in the algorithmic subdomain.

Value

A list of class bco.domain

Examples

empirical <- data.frame(
  "key" = c("false_negative_alignment_hits", "false_discovery"),
  "value" = c("<0.0010", "<0.05"),
  stringsAsFactors = FALSE
)

algorithmic <- data.frame(
  "key" = c("false_positive_mutation_calls", "false_discovery"),
  "value" = c("<0.00005", "0.005"),
  stringsAsFactors = FALSE
)

compose_error(empirical, algorithmic) %>% convert_json()
#> {
#>   "empirical_error": {
#>     "false_negative_alignment_hits": "<0.0010",
#>     "false_discovery": "<0.05"
#>   },
#>   "algorithmic_error": {
#>     "false_positive_mutation_calls": "<0.00005",
#>     "false_discovery": "0.005"
#>   }
#> }