Parse Biocompute Object From JSON File to R Object

Parse Biocompute Object From JSON File to R Object

read_bco(x, ...)

Arguments

x

BioCompute Object .json file

...

Additional parameters for fromJSON.

Value

A list of class bco

Examples

bco <- tempfile(fileext = ".json")
bco <- generate_example("HCV1a") %>%
  convert_json() %>%
  export_json(bco)
bco %>% read_bco()
#> $spec_version
#> [1] "https://w3id.org/biocompute/1.4.2/"
#> 
#> $object_id
#> [1] "https://biocompute.sbgenomics.com/bco/432df52b-575b-431d-bcd7-955f882660e8"
#> 
#> $etag
#> [1] "d18deb41a97a3108e743231af5fec0055dac321b945a569ca58bf33503ec526b"
#> 
#> $provenance_domain
#> $provenance_domain$name
#> [1] "HCV1a ledipasvir resistance SNP detection"
#> 
#> $provenance_domain$version
#> [1] "1.0.0"
#> 
#> $provenance_domain$review
#>     status
#> 1 approved
#> 2 approved
#>                                                           reviewer_comment
#> 1 Approved by [company name] staff. Waiting for approval from FDA Reviewer
#> 2                                               The revised BCO looks fine
#>         date
#> 1 1510507848
#> 2 1513110648
#>                                                                                                       reviewer
#> 1 Jane Doe, Seven Bridges Genomics, example@sevenbridges.com, curatedBy, https://orcid.org/0000-0000-0000-0000
#> 2                                  John Doe, U.S. Food and Drug Administration, example@fda.gov, curatedBy, NA
#> 
#> $provenance_domain$derived_from
#> [1] "https://github.com/biocompute-objects/BCO_Specification/blob/1.2.1-beta/HCV1a.json"
#> 
#> $provenance_domain$obsolete_after
#> [1] "2018-11-12T12:30:48-0500"
#> 
#> $provenance_domain$embargo
#> [1] "2017-10-12T12:30:48-0500" "2017-11-12T12:30:48-0500"
#> 
#> $provenance_domain$created
#> [1] "2017-01-20T09:40:17-0500"
#> 
#> $provenance_domain$modified
#> [1] "2019-05-10T09:40:17-0500"
#> 
#> $provenance_domain$contributors
#>       name                       affiliation                    email
#> 1 Jane Doe            Seven Bridges Genomics example@sevenbridges.com
#> 2 John Doe U.S. Food and Drug Administration          example@fda.gov
#>           contribution                                 orcid
#> 1 createdBy, curatedBy https://orcid.org/0000-0000-0000-0000
#> 2           authoredBy                                    NA
#> 
#> $provenance_domain$license
#> [1] "https://creativecommons.org/licenses/by/4.0/"
#> 
#> 
#> $usability_domain
#> [1] "Identify baseline single nucleotide polymorphisms (SNPs)[SO:0000694], (insertions)[SO:0000667], and (deletions)[SO:0000045] that correlate with reduced (ledipasvir)[pubchem.compound:67505836] antiviral drug efficacy in (Hepatitis C virus subtype 1)[taxonomy:31646]"
#> [2] "Identify treatment emergent amino acid (substitutions)[SO:1000002] that correlate with antiviral drug treatment failure"                                                                                                                                                 
#> [3] "Determine whether the treatment emergent amino acid (substitutions)[SO:1000002] identified correlate with treatment failure involving other drugs against the same virus"                                                                                                
#> 
#> $extension_domain
#> $extension_domain$fhir_extension
#> $extension_domain$fhir_extension$fhir_endpoint
#> [1] "https://fhirtest.uhn.ca/baseDstu3"
#> 
#> $extension_domain$fhir_extension$fhir_version
#> [1] "3"
#> 
#> $extension_domain$fhir_extension$fhir_resources
#> $extension_domain$fhir_extension$fhir_resources[[1]]
#>   fhir_id fhir_resource
#> 1   21376      Sequence
#> 
#> $extension_domain$fhir_extension$fhir_resources[[2]]
#>   fhir_id    fhir_resource
#> 1 6288583 DiagnosticReport
#> 
#> $extension_domain$fhir_extension$fhir_resources[[3]]
#>   fhir_id    fhir_resource
#> 1   25544 ProcedureRequest
#> 
#> $extension_domain$fhir_extension$fhir_resources[[4]]
#>   fhir_id fhir_resource
#> 1   92440   Observation
#> 
#> $extension_domain$fhir_extension$fhir_resources[[5]]
#>   fhir_id       fhir_resource
#> 1 4588936 FamilyMemberHistory
#> 
#> 
#> 
#> $extension_domain$scm_extension
#> $extension_domain$scm_extension$scm_repository
#> [1] "https://github.com/example/repo1"
#> 
#> $extension_domain$scm_extension$scm_type
#> [1] "git"
#> 
#> $extension_domain$scm_extension$scm_commit
#> [1] "c9ffea0b60fa3bcf8e138af7c99ca141a6b8fb21"
#> 
#> $extension_domain$scm_extension$scm_path
#> [1] "workflow/hive-viral-mutation-detection.cwl"
#> 
#> $extension_domain$scm_extension$scm_preview
#> [1] "https://github.com/example/repo1/blob/c9ffea0b60fa3bcf8e138af7c99ca141a6b8fb21/workflow/hive-viral-mutation-detection.cwl"
#> 
#> 
#> 
#> $description_domain
#> $description_domain$keywords
#> [1] "HCV1a"                    "Ledipasvir"              
#> [3] "antiviral resistance"     "SNP"                     
#> [5] "amino acid substitutions"
#> 
#> $description_domain$xref
#>          namespace              name
#> 1 pubchem.compound  PubChem-compound
#> 2           pubmed            PubMed
#> 3               so Sequence Ontology
#> 4         taxonomy          Taxonomy
#>                                             ids              access_time
#> 1                                      67505836 2017-01-20T09:40:17-0500
#> 2                                      26508693 2017-01-21T09:40:17-0500
#> 3 SO:000002, SO:0000694, SO:0000667, SO:0000045 2017-01-22T09:40:17-0500
#> 4                                         31646 2017-01-23T09:40:17-0500
#> 
#> $description_domain$platform
#> [1] "Seven Bridges Platform"
#> 
#> $description_domain$pipeline_steps
#>   step_number         name                               description version
#> 1           1 HIVE-hexagon Alignment of reads to a set of references     1.3
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         prerequisite
#> 1 Hepatitis C virus genotype 1, Hepatitis C virus type 1b complete genome, Hepatitis C virus (isolate JFH-1) genomic RNA, Hepatitis C virus clone J8CF, complete genome, Hepatitis C virus S52 polyprotein gene, https://www.ncbi.nlm.nih.gov/nuccore/22129792, https://www.ncbi.nlm.nih.gov/nuccore/5420376, https://www.ncbi.nlm.nih.gov/nuccore/13122261, https://www.ncbi.nlm.nih.gov/nuccore/386646758, https://www.ncbi.nlm.nih.gov/nuccore/295311559, 2017-01-24 09:40:17, 2017-01-24 09:40:17, 2017-01-24 09:40:17, 2017-01-24 09:40:17, 2017-01-24 09:40:17
#>                                                                                                                                         input_list
#> 1 https://example.com/dna.cgi?cmd=objFile&ids=514683, https://example.com/dna.cgi?cmd=objFile&ids=514682, 2017-01-24 09:40:17, 2017-01-24 09:40:17
#>                                                                                                                                       output_list
#> 1 https://example.com/data/514769/allCount-aligned.csv, https://example.com/data/514801/SNPProfile*.csv, 2017-01-24 09:40:17, 2017-01-24 09:40:17
#> 
#> 
#> $execution_domain
#> $execution_domain$script
#> [1] "https://example.com/workflows/antiviral_resistance_detection_hive.py"
#> 
#> $execution_domain$script_driver
#> [1] "shell"
#> 
#> $execution_domain$software_prerequisites
#>            name      version
#> 1  HIVE-hexagon babajanian.1
#> 2 HIVE-heptagon   albinoni.2
#>                                                                                                                    uri
#> 1 https://example.com/dna.cgi?cmd=dna-hexagon&cmdMode=-, 2017-01-24 09:40:17, d60f506cddac09e9e816531e7905ca1ca6641e3c
#> 2                                      https://example.com/dna.cgi?cmd=dna-heptagon&cmdMode=-, 2017-01-24 09:40:17, NA
#> 
#> $execution_domain$external_data_endpoints
#> $execution_domain$external_data_endpoints[[1]]
#>           name                                     url
#> 1 generic name protocol://domain:port/application/path
#> 
#> $execution_domain$external_data_endpoints[[2]]
#>                   name                        url
#> 1 access to ftp server ftp://data.example.com:21/
#> 
#> $execution_domain$external_data_endpoints[[3]]
#>                            name                                           url
#> 1 access to e-utils web service https://eutils.ncbi.nlm.nih.gov/entrez/eutils
#> 
#> 
#> $execution_domain$environment_variables
#> $execution_domain$environment_variables$HOSTTYPE
#> [1] "x86_64-linux"
#> 
#> $execution_domain$environment_variables$EDITOR
#> [1] "vim"
#> 
#> 
#> 
#> $parametric_domain
#>                          param value step
#> 1                         seed    14    1
#> 2            minimum_match_len    66    1
#> 3 divergence_threshold_percent  0.30    1
#> 4             minimum_coverage    15    2
#> 5                  freq_cutoff  0.10    2
#> 
#> $io_domain
#> $io_domain$input_subdomain
#>                                                                                                                 uri
#> 1             Hepatitis C virus genotype 1, https://www.ncbi.nlm.nih.gov/nuccore/22129792, 2017-01-24T09:40:17-0500
#> 2 Hepatitis C virus type 1b complete genome, https://www.ncbi.nlm.nih.gov/nuccore/5420376, 2017-01-24T09:40:17-0500
#> 
#> $io_domain$output_subdomain
#>   mediatype
#> 1  text/csv
#> 2  text/csv
#>                                                                               uri
#> 1 https://example.com/data/514769/dnaAccessionBased.csv, 2017-01-24T09:40:17-0500
#> 2       https://example.com/data/514801/SNPProfile*.csv, 2017-01-24T09:40:17-0500
#> 
#> 
#> $error_domain
#> $error_domain$empirical_error
#> $error_domain$empirical_error$false_negative_alignment_hits
#> [1] "<0.0010"
#> 
#> $error_domain$empirical_error$false_discovery
#> [1] "<0.05"
#> 
#> 
#> $error_domain$algorithmic_error
#> $error_domain$algorithmic_error$false_positive_mutation_calls
#> [1] "<0.00005"
#> 
#> $error_domain$algorithmic_error$false_discovery
#> [1] "0.005"
#> 
#> 
#> 
#> attr(,"class")
#> [1] "list" "bco"