Compose BioCompute Object - Execution Domain (v1.3.0)

Compose BioCompute Object - Execution Domain (v1.3.0)

compose_execution_v1.3.0(script = NULL, script_driver = NULL,
  software_prerequisites = NULL, external_data_endpoints = NULL,
  environment_variables = NULL)

compose_execution(script = NULL, script_driver = NULL,
  software_prerequisites = NULL, external_data_endpoints = NULL,
  environment_variables = NULL)

Arguments

script
Character string. Points to internal or external references to an object that was used to perform computations for this BCO instance.

script_driver
Character string. Indicate what kind of executable can be launched in order to perform a sequence of commands described in the script in order to run the pipeline.

software_prerequisites
Data frame. The minimal necessary prerequisites, library, and tool versions needed to successfully run the script to produce BCO. Variables include name, version, uri, access_time, and sha1_chksum. Each row is one item in the output subdomain.

external_data_endpoints
Data frame. The minimal necessary domain-specific external data source access to successfully run the script to produce the BCO. Variables include mediatype, name, and url. Each row is one item in the output subdomain.

environment_variables
Data frame. Key-value pairs useful to configure the execution environment on the target platform. Variables include key and value.

Examples

script <- "https://example.com/workflows/antiviral_resistance_detection_hive.py"
script_driver <- "shell"
software_prerequisites <- data.frame(
  "name" = c("HIVE-hexagon", "HIVE-heptagon"),
  "version" = c("babajanian.1", "albinoni.2"),
  "uri" = c(
    "https://example.com/dna.cgi?cmd=dna-hexagon&cmdMode=-",
    "https://example.com/dna.cgi?cmd=dna-heptagon&cmdMode=-"
  ),
  "access_time" = c(
    as.POSIXct("2017-01-24T09:40:17", format = "%Y-%m-%dT%H:%M:%S", tz = "EST"),
    as.POSIXct("2017-01-24T09:40:17", format = "%Y-%m-%dT%H:%M:%S", tz = "EST")
  ),
  "sha1_chksum" = c("d60f506cddac09e9e816531e7905ca1ca6641e3c", NA),
  stringsAsFactors = FALSE
)
external_data_endpoints <- data.frame(
  "name" = c("generic name", "access to ftp server", "access to e-utils web service"),
  "url" = c(
    "protocol://domain:port/application/path",
    "ftp://data.example.com:21/",
    "https://eutils.ncbi.nlm.nih.gov/entrez/eutils"
  ),
  stringsAsFactors = FALSE
)
environment_variables <- data.frame(
  "key" = c("HOSTTYPE", "EDITOR"),
  "value" = c("x86_64-linux", "vim")
)

compose_execution(
  script, script_driver, software_prerequisites, external_data_endpoints, environment_variables
) %>% convert_json()

#> {
#>   "script": "https://example.com/workflows/antiviral_resistance_detection_hive.py",
#>   "script_driver": "shell",
#>   "software_prerequisites": [
#>     {
#>       "name": "HIVE-hexagon",
#>       "version": "babajanian.1",
#>       "uri": [
#>         {
#>           "uri": "https://example.com/dna.cgi?cmd=dna-hexagon&cmdMode=-",
#>           "access_time": "2017-01-24 09:40:17",
#>           "sha1_chksum": "d60f506cddac09e9e816531e7905ca1ca6641e3c"
#>         }
#>       ]
#>     },
#>     {
#>       "name": "HIVE-heptagon",
#>       "version": "albinoni.2",
#>       "uri": [
#>         {
#>           "uri": "https://example.com/dna.cgi?cmd=dna-heptagon&cmdMode=-",
#>           "access_time": "2017-01-24 09:40:17",
#>           "sha1_chksum": "NA"
#>         }
#>       ]
#>     }
#>   ],
#>   "external_data_endpoints": [
#>     [
#>       {
#>         "name": "generic name",
#>         "url": "protocol://domain:port/application/path"
#>       }
#>     ],
#>     [
#>       {
#>         "name": "access to ftp server",
#>         "url": "ftp://data.example.com:21/"
#>       }
#>     ],
#>     [
#>       {
#>         "name": "access to e-utils web service",
#>         "url": "https://eutils.ncbi.nlm.nih.gov/entrez/eutils"
#>       }
#>     ]
#>   ],
#>   "environment_variables": {
#>     "HOSTTYPE": "x86_64-linux",
#>     "EDITOR": "vim"
#>   }
#> }